cleanup money (#3535)
* restructures ResultType and decouples printing into ResultPrinters:
  - ResultType is now very simple mix of primitive enum and LIST. (We hope to remove even more logic from it)
  - Result printing was always weird with Mappings embedded in StringT. We've moved that responsibility into ResultInfo via SelectResultInfo for mapped Selects.
  - This PR opens the possibility to embed anonymization into the ResultInfo.
  - PrintSettings now caches the C10n calls as they were before that sometimes invoked per cell, which should make rendering faster.
* Reworks Printers:
  - We inject the result specific Printers using the printerFactory. This untangles the burden of formatting.
  - Now a Select claiming to produce a specific type only needs to guarantee they are in a format capable of being handled by the printers it elects to use: This allows us later to make a printer that alters the result before printing it, without having to inject into all renderers.

* reworks Money to no longer be "cent" based and instead just use it as storage primitive. Is now BigDecimal
  - pull default for decimalShift from CurrencyConfig

+++ backend/src/main/java/com/bakdata/conquery/models/events/Bucket.java
@@ -162,7 +162,7 @@ public class Bucket extends IdentifiableImpl<BucketId> implements NamespacedIden
 		return ((DecimalStore) getStore(column)).getDecimal(event);
 	}
 
-	public long getMoney(int event, @NotNull Column column) {
+	public BigDecimal getMoney(int event, @NotNull Column column) {
 		return ((MoneyStore) getStore(column)).getMoney(event);
 	}
